requirements.txt
django==2.2.15
django-extensions==3.0.5
django-model-utils==4.0.0
